[core] add benchmark for vector tile parsing
